+++ example/src/docs/recipes/blog/blog_collection.tsx
@@ -8,7 +8,7 @@ export const blogCollection = buildCollection<BlogEntry>({
     views: [{
         path: "preview",
         name: "Preview",
-        builder: (props) => <BlogEntryPreview {...props}/>
+        Builder: BlogEntryPreview
     }],
     properties: {
         name: buildProperty({