+++ lib/src/core/components/EntityCollectionTable/internal/EntityCollectionRowActions.tsx
@@ -166,7 +166,7 @@ export function EntityCollectionRowActions<M extends Record<string, any>>({
                         anchorEl={anchorEl}
                         open={Boolean(anchorEl)}
                         onClose={closeMenu}
-                        elevation={2}
+                        elevation={3}
                     >
                         {deleteEnabled && <MenuItem onClick={onDeleteClick}>
                             <ListItemIcon>