+++ lib/src/core/theme.ts
@@ -170,10 +170,7 @@ export const createCMSDefaultTheme = (
         ...original,
         // @ts-ignore
         shadows: original.shadows.map((shadow, index) => {
-            // if (index === 1) {
-            //     return "0px 2px 4px -1px rgba(0,0,0,0.2),0px 4px 5px 0px rgba(0,0,0,0.14),0px 1px 10px 0px rgba(0,0,0,0.12)";
-            // }
-            if (index === 2) {
+            if (index === 3) {
                 return "0px 0px 0px 0px rgba(0,0,0,0.18), 0px 0px 10px 0px rgba(0,0,0,0.12), 0px 0px 14px 2px rgba(0,0,0,0.10)"
             }
             return shadow;