+++ lib/src/test/components/util.ts
@@ -7,7 +7,7 @@ export const productCollection: EntityCollection = {
         {
             path: "custom_view",
             name: "Test custom view",
-            builder: ({}) => undefined
+            Builder: ({}) => null
         }
     ],
     properties: {